py/requirements.txt
eel==0.14.0
requests==2.27.1
opencv-python
Pillow==9.1.1
numpy==1.26.0